+++ apps/www/src/lib/registry/new-york/ui/form/FormItem.vue
@@ -1,10 +1,3 @@
-<script lang="ts">
-import { type InjectionKey } from 'vue'
-
-export const FORM_ITEM_INJECTION_KEY
-  = Symbol() as InjectionKey<string>
-</script>
-
 <script lang="ts" setup>
 import { provide } from 'vue'
 import { useId } from 'radix-vue'
