+++ apps/www/src/lib/registry/new-york/ui/form/useFormField.ts
@@ -1,6 +1,6 @@
 import { FieldContextKey, useFieldError, useIsFieldDirty, useIsFieldTouched, useIsFieldValid } from 'vee-validate'
 import { inject } from 'vue'
-import { FORM_ITEM_INJECTION_KEY } from './FormItem.vue'
+import { FORM_ITEM_INJECTION_KEY } from './InjectionKey'
 
 export function useFormField() {
   const fieldContext = inject(FieldContextKey)
